requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
webdriver-manager>=4.0.0
pandas>=2.1.0
jinja2>=3.1.0
python-dotenv>=1.0.0
schedule>=1.2.0
